FlexibleSettings, settings -> parameters, deprecated api
