+++ src/main/java/io/github/eocqrs/kafka/consumer/KfConsumer.java
@@ -41,6 +41,7 @@ import java.util.List;
  * @author Aliaksei Bialiauski (abialiauski.dev@gmail.com)
  * @since 0.0.0
  */
+@SuppressWarnings("removal")
 public final class KfConsumer<K, X> implements Consumer<K, X> {
 
   /**
