+++ src/test/java/io/github/eocqrs/kafka/parameters/KfFlexibleSettingsTest.java
@@ -1,0 +1,44 @@
+package io.github.eocqrs.kafka.parameters;
+
+import io.github.eocqrs.kafka.consumer.settings.KfConsumerParams;
+import org.apache.kafka.clients.consumer.KafkaConsumer;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+/**
+ * Test case for {@link KfFlexibleSettings}.
+ *
+ * @author Ivan Ivanchuck (l3r8y@duck.com)
+ * @since 0.0.2
+ */
+final class KfFlexibleSettingsTest {
+
+  @Test
+  void createsConsumerWithCustomSettings() {
+    try (
+      final KafkaConsumer<Object, Object> consumer =
+        new KfFlexibleSettings<>(
+          new KfConsumerParams(
+            new KfParams(
+              new BootstrapServers("localhost:9092"),
+              new GroupId("1"),
+              new KeyDeserializer("org.apache.kafka.common.serialization.StringDeserializer"),
+              new ValueDeserializer("org.apache.kafka.common.serialization.StringDeserializer"),
+              new KfFlexibleSettingsTest.HeartbeatIntervalMs("4234")
+            )
+          )
+        ).consumer()
+      ) {
+      Assertions.assertThat(consumer).isNotNull();
+      assertDoesNotThrow(consumer::subscription);
+    }
+  }
+
+  private static final class HeartbeatIntervalMs extends KfAttrEnvelope {
+    HeartbeatIntervalMs(final String value) {
+      super(value, "heartbeat.interval.ms");
+    }
+  }
+}
