+++ src/test/java/io/github/eocqrs/kafka/parameters/KfParamsTest.java
@@ -20,41 +20,31 @@
  * SOFTWARE.
  */
 
-package io.github.eocqrs.kafka.settings;
+package io.github.eocqrs.kafka.parameters;
 
-import io.github.eocqrs.kafka.ParamsAttr;
-import io.github.eocqrs.kafka.xml.NameInCamelCase;
-import org.cactoos.text.FormattedText;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Test;
 
 /**
- * It's a wrapper for a string value that can be converted to XML.
+ * Test case for {@link KfParamsTest}.
  *
  * @author Ivan Ivanchuk (l3r8y@duck.com)
  * @since 0.0.2
  */
-public abstract class AttrEnvelope implements ParamsAttr {
+final class KfParamsTest {
 
-  protected final String value;
-
-  protected final String name;
-
-  protected AttrEnvelope(final String value, final String name) {
-    this.value = value;
-    this.name = name;
-  }
-
-  @Override
-  public final String asXml() {
-    return new FormattedText(
-      "<%s>%s</%s>",
-      new NameInCamelCase(this.name),
-      this.value,
-      new NameInCamelCase(this.name)
-    ).toString();
-  }
-
-  @Override
-  public final String name() {
-    return this.name;
+  @Test
+  void representsXmlCorrectly() {
+    MatcherAssert.assertThat(
+      "Contains right tags",
+      new KfParams(
+        new GroupId("103"),
+        new KeySerializer("ks")
+      ).asXml(),
+      Matchers.equalTo(
+        "<groupId>103</groupId>\n<keySerializer>ks</keySerializer>"
+      )
+    );
   }
 }
